requirements.txt
flake8
wikipedia
google-api-python-client
discord-py==v2.1.0
urllib3>=1.24.2
python-dateutil
youtube-dl>=2019.09.01
PyNaCl>=1.3.0
boto3>=1.12.0
psycopg2
psutil
